+++ CHANGELOG.md
@@ -12,7 +12,8 @@ _general description_
 - **added** - `<concepts>` and `<ranges>` check and import
 - **added** - simple logging system with different levels and color support
 - **added** - unit test framework without macros
-- **added** - time util
+- **added** - time clock implementation
+- **added** - engine base and fixed update loop
 
 ---